Major changes to the processing in the script.
The script now imports and exports to .html formats. Functions have been added to parse html files by stripping out everything but the table element, removing all but the required information, and then outputting that information to a custom PS Object.

For the file export, a for loop has been introduced to search through each table row. If any of the table rows contain the string Online or Offline, conditional formatting will be applied using green and red, respectively.